Comments - Defect Number: 3160; Defect RecordID: 3156; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Details > Commission Tab New Functionality; Defect Description: A recent change for some Logic20/20 employees (Client Directors and Senior Managers) has been the implementation of targets for sales and delivery of client projects. Bonuses are now based on the amount they contribute to the sales and delivery of our projects. The Finance team has been tracking this manually right now, but we recognize that adding some of this tracking functionality to PM would make help make the process more efficient.
Attached is a mock-up of the enhancements I am requesting to the current "Commissions" tab in Project Details. All current functionality on that tab should be deprecated. Access to the tab should be limited to individuals with the title of Senior Manager and the Senior Leadership and Admin roles.

The UI on the Project Details will feed an Excel-based extract (see mock-up also attached) that should be added as a new tab entitled "Sales Delivery Attributions" to the Attainment report. Please also create a stored procedure so that this report could be run separately if a person wished. Please set up the stored procedure the same as the others that you've done for us recently, allowing us to specify the time range.

Unrelated to the Commission tab, but related to general clean up and organization of the information on the Project Tab, please move the "Project Time Types" tab and related functionality, and move it up into the main row of tabs. It currently sits as an "orphan" at the bottom of the Project Details page, and I've received user feedback that it's placement on the page can cause confusion as to its purpose.

Could you please review the attached and let me know of any questions? What will be the approximate level of effort for this request?