Comments - Defect Number: 3170; Defect RecordID: 3166; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Changing Password Business Rule > Allowing New Hires to change password from temporary to permanent within 24 hours; Defect Description: IT alerted me that they have encountered an issue that new hires are not able to change their PM passwords on their first day of work. The new hires receive an error message after signing in with their temporary password and attempting to change it to a permanent password saying that the system does not allow more than once password change within a 24 hour period.
Because of the recent request change to the New Hire workflow, new PM accounts and email notifications are now sent out at midnight of the person's first day. As per the current rule, PM will only allow a password to be changed once per day and our hypothesis as to why the new hires are receiving this error message is that the system considers the assignment of the temporary password on hire date as the first password change and thus will not allow another one within a 24 hour time period.

Can we make a small change to the current business rules so that a user can change his/her password within 24 hours of receiving the initial "Welcome to Practice Management" email and temporary password? In all other cases, the current rule that a password cannot be change more than once within a 24 hour time period should remain in effect.