Comments - Defect Number: 3157; Defect RecordID: 3153; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Details > Add resource selection enhancement; Defect Description: As reflected in Hotfix #3156, we struggled to correct a data entry error when a strawman was not replaced with a named person because that person was a terminated employee. In order for us to be able to correct these types of data entry errors in the future, can we implement the following change:
When adding a resource in "Project Details", the drop-down list should contain the names of all consultants (i.e., those who do not = IsInternal) who had an active status during the specified time period even if they have a status of "terminated" today. However, business logic should be applied so that the system will not assign a project end date for a terminated employee later than the individual's termination date.

For example, the data entry error articulated in Hotfix #3156 was to replace the Strawman in the "March Milestone" of Project P000170 with Jeremy McCarrell, a now terminated employee who was an active employee with the company from 3/1/2013 to 3/20/2013 (he terminated with the company on 3/20/2013). If I were to add a new resource to this milestone today, I expect to see Jeremy McCarrell's name in the picklist to select, as he was an active employee during that March time period. However, if I tried to select his end date on project for a date later than 3/20/2013, I would expect the system to give me an error message saying that the resource cannot be added to the project past <insert date> due to his/her termination from the company.