Comments - Defect Number: 3173; Defect RecordID: 3169; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports > Create three Exception Reports; Defect Description: The Finance Team has requested a set of three Exception reports that will help them track certain undesirable behaviors that could lead to a loss in revenue by not charging all the hours to the client that may have been possible. The team's current process is that they discover this information during the invoicing process, but by that time it is too late to be able to correct the situation. With regular exception reporting, they will be able to proactively identify potential issues and work with the project leads to rectify if appropriate.
There are two deliverables for this request:
1. Create a stored procedure for each of the three reports so that they can be run on an ad hoc basis. Like you have been doing for other stored procedures, please make the query flexible so that a particular date range can be returned.
2. Every Monday morning, the system should automatically send an Excel export containing the three exception reports (one on each tab) to Finance@logic2020.com.
a. The time range for this extract should be for the previous week (Sunday  Sunday).
b. Title the email Practice Management Project and Time Entry Exception Reports for the Week of <Insert Start Date> to <Insert End Date>

Exception Report Descriptions:
1. $0 Hourly Rate: This report displays the people who have been allocated to a milestone with a Revenue type of Hourly) with an hourly rate of $0.
a. A current example of this is P000222 where Pavel Lyubich is assigned to Milestone Core team with a $0 bill rate for the period of 7/30/2013  8/30/2013
b. Mock-up is attached, with a description of each requested field

2. Unassigned Resources Charging to a Project: This report displays employees who are charging Billable and Non-Billable hours to a project that they have not been assigned to.
a. Mock-up is attached, with a description of each requested field

[Harshitha]
In "Unassigned Resources Charging to a Project" report, do we need to display the person who has time entries for complete week? or display the person if he has atleast one time entry for that project in that week?

LN: The latter. Any time entries made to the project - for one or more days - should be included in the report.

3. Assigned Resources Not Charging to a Project: This report displays the opposite of the above: it displays a list of employees who were assigned to a project milestone but did not charge any billable hours to it within the specified time period.
a. Mock-up is attached, with a description of each requested field

[Harshitha]
In "Assigned Resources Not Charging to a Project" report, do we need to display the person who does not have time entries for complete weekfor that project? or display the person if he has atleast one day with out time entry in that week?

LN: Yes, if the person should have billed the project at least one day of the five business days, but did not, then the person should be included in the report.