Comments - Defect Number: 3177; Defect RecordID: 3173; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Detail > Resource selection does not include Client Directors/Practice Directors; Defect Description: It was my mistake that when we implemented Ticket #3157 I neglected to state the requirement that Client Directors and Practice Directors should remain in the drop-down list and remain selectable to add/edit to a project.
As a result, David Barnes is trying to add himself to P000222 and could not locate himself in the drop-down list. David's title is Client Director. As a member of the Senior Leadership team, he will sometimes charge for his services to a client, and needs to be able to add himself to a project to properly calculate revenue/margin/COGS. Not being able to add himself is thus now leading to incorrect financial forecasting.

Client Directors (Tim Blank, Nick Maddock, David Barnes) and Practice Director (Karl Haberl) are in the "Administration" Practice Area but in the Consulting division. Can we please update the business logic for the resource drop list to include these individuals as soon as possible? Please update and push to Production immediately.