Comments - Defect Number: 3182; Defect RecordID: 3178; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Detail > Resource selection no longer include Strawmen; Defect Description: Another bug with our recent 3157 fix was discovered today. It was again my fault not to specifically state that Strawmen (isStrawman=yes) should be included in the resource details drop-down, and during my testing, I did not catch that those resources were no longer added. Strawmen are used by our organization to model potential projects and therefore be able to correctly
Please add Strawman resources back into the resource selection drop down (push immediately into Production).

I believe (as please conduct your assessment and give me your recommendations as well) that the two hotfix issues we have encountered in the past week are both resulted from the changes to the business rules surrounding Practice Area and department that we applied when we rolled out this new functionality. As soon as the short-term fix of putting back Strawmen into the resource drop-down in complete, can you please roll back any changes that we made to the business logic of who should be included in the list and align it back to the old business rules with the exception of the terminated resources enhancement that was the central pillar to the original request. I still want to retain the functionality of being able to select currently terminated resources on the project who were active during the project milestones. This component of my ask is not a "hot fix" request - i.e., doesn't have to be completed tomorrow.