Comments - Defect Number: 3183; Defect RecordID: 3179; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports > Time by Project Report; Defect Description: Projected Hours do not appear to be calculating correctly on the Time by Project report. When I was testing the new report's functionality in UAT, my focus was on the new "Hourly Rate" and "Est. Billings" columns and so did not catch this apparent miscalculation. The "Projected Hours" calculation should not have been impacted by the addition of these two new columns.
Please see the screenshots below for my analysis and thinking. Could you please investigate?

For the screenshot below: Each resource is billing the same amount of hours on the project per day. Even allowing for potential PTO, there is a too wide of a difference between each person's projected hours, leading me to believe that this calculation is not correct.



For the screenshot below: When pulling the By Person Time Report for an individual on this team, his projected hours are different than stated above. For nine working days in this previous period, 72 hours is what I expect to see for the person's Projected Hours.