Comments - Defect Number: 3178; Defect RecordID: 3174; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports > Account > Open Access to individuals with title of "Manager"; Defect Description: My understanding is that the "Account" report (Reports > Account) is only accessible to the Admin role and to individuals with the title of "Client Director" and "Senior Manager". We have a request to open permission to this report to individuals with the title of "Manager" as well.
Could you please a) confirm what titles/roles/seniority can see this report and b) open permissions to the Manager level?

[Harshitha]
We are showing Account report basing on the role of user. Currently we are showing Account report only for the users who are with at least one of the roles: Administrator, Client Director, Business Unit Manager, and Senior Leadership.

Attached document contain the currently active users with Manager Title and their roles, Could you please take a look and let us know for which other roles do we need to show the Account report?