Comments - Defect Number: 3179; Defect RecordID: 3175; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Summary Report > Projects not appearing in view ; Defect Description: Employee Patrick O'Briant reports an issue in viewing projects in the Projects > Summary view. He reports that when he runs Project Summary (using filters of "Microsoft" account and "All" for the rest of the filters) projects that have Stephanie Stocker as a salesperson do not appear for him. His hypothesis is that because Stephanie is a terminated employee, Practice Management is filtering out projects assigned to her as a sales person. When I run the report, I can see all projects, including ones where Stephanie Stocker is a salesperson, as I expect, but when I become Patrick as a user, I was able to confirm his issue.
Could you please investigate any potential cause of why Patrick, based on his title/roles/seniority allocated to him, may be seeing a different view than me in regard to projects assigned to Stephanie Stocker?


[Harshitha]
In the project summary page, if an Administrator user selects 'All' in sales persons list then we show the projects of terminated sales persons also, But if the users with role other than administrator selects All in sales persons list then we show the projects of only active sales persons.

This is the reason why Lisa is able to see the projects of Stephanie Stocker and Patrick O'Briant is not able to.

This is behaving in the same way from the start.

We have an option to resolve this:
1. Showing terminated sales persons in the drop down of project summary page such that the logged in user selects the sales person whose projects he wish to view.

LN: Thank you very much for the analysis. Please proceed with your recommendation.

---

A completely unrelated request: when I submit a new ticket, TTP automatically populates the assignment to Madhan. Could you please instruct me how to change it to your name?

[Harshitha]
We have updated the configuration of Test track such that whenever a ticket is created then it will be assigned to Harshitha Tenneti instead of Madhan.
Attached a document of steps for this task.

LN: Thank you very much!