Comments - Defect Number: 3191; Defect RecordID: 3187; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports > Payroll Distribution Report > Change delivery dates; Defect Description: The Payroll Distribution Report is currently automatically sent by the system to Finance@logic2020.com on the 3rd and 18th of every month. Can we please change the dates so that report runs and is sent to Finance on the 4th and 19th of every month?
The Finance team is experiencing a lot of flux in the hours charged by consultants in the first three days after the end of a payroll period, leading to the Distribution report being inaccurate and causing them to ask me to run the stored procedure again. By postponing the running of the report by one day, the Finance team anticipates that most time entry errors/changes will have been completed, eliminating the need for duplicative work.