Comments - Defect Number: 3135; Defect RecordID: 3131; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Consulting Utilization Report > W2-Hourly employees unassigned to a project not appearing in report; Defect Description: Employee "Joshua Weaver" transitioned from a "Pay Type" of "W2-Salary" to "W2-Hourly" status effective today, 4/16/13. He is currently unassigned to a "Active," "Projected," or "Internal" project type. When the Consulting Utilization Report is run using the default filters, Josh does not appear in the results. He only appears when the "Experimental" project type filter is checked.
Employees with a "pay type" of W2-Hourly should appear in the Consulting Utilization Report in the same manner as all other resources. In this case, Josh Weaver should be listed and marked with 0% utilization. Could you please investigate why this employee was not appearing in the graph and resolve accordingly?