Comments - Defect Number: 3184; Defect RecordID: 3180; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project>Milestone detail>resources tab> Start date is not being updated on hire date change.; Defect Description: Scenario 1: In Milestone detail page the resources start date is not being updated on hire date change. 1. Add a person with hire date as 8/1/2013 2. Now add this persons as a resource to a milestone with dates 8/1/2013-8/31/2013 3. Update the hire date of the person to be 8/10/2013.
Actual: In the above scenario we expect the resources dates to be updated to 8/10/2013-8/31/2013, but the hire date is not being updated to 8/10/2013 and still remaining at 8/1/2013.

Expected: Please update it in such a way that the resources start date is set to 8/10/2013 when hire date is changed to 8/10/2013.


Scenario 2:
Message is not being displayed when a person in milestone resources list & the hire date is moved prior to current hire date.
1. Let us assume that there is a person with hire date as 8/1/2013.
2. Add that person to milestone with dates 8/1/2013-8/31/2013
3. Update the persons hire date to be 7/20/2013

Actual: No message is being displayed to the user asking to update the milestone resources dates accordingly if needed

Expected: Message should be displayed to the user asking to update the milestone resources dates accordingly if needed