Comments - Defect Number: 3200; Defect RecordID: 3196; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Payroll Distribution Report for Calendar 2010; Defect Description: The Payroll Distribution Report that the team created has been enormously beneficial to the Finance team, and is used for a variety of purposes. They are working on a project right now in which they would like all data from the system since the inception of the system in the Payroll Distribution format. Using the stored procedure created, I was able to create calendar year 2012 and calendar year 2011 reports, but encountered an error message when I tried to run the report for calendar year 2010:
Msg 8134, Level 16, State 1, Procedure PayRollDistribution, Line 23
Divide by zero error encountered.
Warning: Null value is eliminated by an aggregate or other SET operation.

I suspect that the cause of this issue is that in 2010, we didn't store Time Entry data with the same parameters as today; the Time Entry system underwent a large enhancement at the beginning of 2012. To that end, I suspect but cannot confirm that we could not pull a Payroll Distribution Report for calendar year 2010 because we don't have the data in the right format to support it.

To that end, I ask for your help in determining whether we have the basic information in the system to help the Finance team with their request: for the calendar year of 2010, they require a list of employees, the Accounts that they worked on, the number of hours worked at that account, and each employee's payrate. A mock-up is below:


Employee Id Employee Practice Area Division Pay Type Is Offshore? Account Name Total Hours Worked on Account Pay Rate in 2010
(Note: Latest if Multiple)
C030901 Neel, Lisa Administration Operations W2-Salary No Microsoft 1000 22

Does the data exist to be able to manipulate it into this type of report? Could you please investigate and present your recommendations?