Comments - Defect Number: 3199; Defect RecordID: 3195; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Account Details > Changing Discount % in Account Details does not filter to Project Details; Defect Description: When a user changes a Default Discount in Account Details (e.g., Brooks Sports), it appears that that change does not filter to an Active project's Project Details. As a result, the Financial Summary tab in Project Details reflects the old Discount Rate, leading to the financial projection to be off.
The specific scenario: The default discount in Production for Brooks Sports was changed from 10% to 0%. The user making the change assumed that any Active project under that account would automatically reflect the new discount rate. For Active project P000290, the original 10% is still reflected.

Could you please help me understand how the code is designed right now? My expectation is that upon making a change to default discount in the Account details, the change would get replicated throughout any Projected, Active, Experimental, or Inactive projects under that account. The change would not be replicated to any projects with a status of "Completed" because we want to preserve the accuracy of the financials for any projects that have been completed and we have already billed for.