Comments - Defect Number: 3203; Defect RecordID: 3199; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Blank alert messages are being shown in system in few scenarios.; Defect Description: We are seeing blank alert messages (see image below) in system in below mentioned few scenarios. Please update them to display correctly. 1. Project details/Project time types tab: If the user adds a work type withempty name then blank alert message is shown. a.Log on to practice management b. Select a project and navigate to that project c.In Project time types tab, add a work type with empty name. d. Click on save
LN: Please prevent the system from allowing a work type to be created that has no characters.

2. Project details/Expenses tab: When the two expenses dates overlap then the error message displayed is blank
a. Log on to practice management
b. Select a project and navigate to that project
c. In expenses tab, add expense dates such that the dates of expenses overlap with other expense.
d. Click on Save button

LN: I would expect that the system would allow the start/end dates of separate expenses to overlap with each other. Could you please adjust the business logic to permit dates to overlap?

3. Project details/expenses tab: if start date and end date of expense are not within the project period range then empty message will be displayed
a. Log on to practice management
b. Select a project and navigate to that project
c. In expenses tab, add expense dates such that they do not match the project dates.
d. Click on add expense button.

LN: The error message should read: "You have attempted to add an expense that was for a period of time outside of the start and end dates for the project. Please readjust the start and end dates for the expense line item."