Comments - Defect Number: 3204; Defect RecordID: 3200; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project cloning with active status; Defect Description: Currently in system we do not allow projects to be set to Active status if there is no milestone added to it, but we are able to set a project to active status without milestone with the following scenario:
Steps to produce:
1. Log on to practice management
2. Add a project without milestone.
3. Clone the project with active status for new project

In the above scenario though the project is cloned with active status without a milestone, a validation message will be displayed up on clikcing the save button of cloned project.
But as this project is with active status then it will be displayed in time entry page projects list.

Could you please let us know if this scenario is expected?