Comments - Defect Number: 3208; Defect RecordID: 3204; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Validation message display while updating person status; Defect Description: While testing 3207 we have observed that the persons with role recruiter cannot update the persons status to Active/Terminated. Also we observed that only the persons with Admin and HR roles can update the status of the person in person detail page.
But the message displayed when recruiter tries to change status is Only administrator can set a status to Active or Terminated.


Please let us know if we need to change the message to read Only administrator or HR person can set a status to Active or terminated.
Or
Please let us know if we need to remove the permission for person with HR role to restrict him from changing persons status.