Comments - Defect Number: 3205; Defect RecordID: 3201; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Hidden margin value in project summary page; Defect Description: Currently a person will not be allowed to see margin of a project in project summary page if at least one resource in that project has seniority is greater than/equal to/next to equal of logged in users seniority.
In the same way the logged in user will not be allowed to view the margin of a resource in project details/milestone details page if the resources seniority greater than/equal to/next to equal of logged in users seniority.

This is a business rule to restrict the user from viewing the margin for persons who has seniority greater than/equal to/next to equal of logged in users seniority.
But the user will be able to view the projects revenue and Margin from group by page.

Steps to reproduce:
1. Log in to Practice management with your credentials.
2. Go to configuration>Troubleshooting> Become userbecome as Annabemoju, Madhan(Updated his person detail accordingly as role=Business Unit Manager, Seniority=Admin)
3. Navigate to project summary page and view WWLP Global Promotions SharePoint Support project.
4.The Margin value will be shown as hidden (As I added myself as resource for that project with role=Administrator and seniority=Admin which is equal to madhans seniority).
5. Now verify the projects margin in Group by page. The margin will be shown to the user.
Could you please let us know your response on this?