Comments - Defect Number: 3213; Defect RecordID: 3209; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: CSAT: Addition of a "Not Applicable" field in CSAT Rating/Score field; Defect Description: We have been encountering occasions where a CSAT score cannot be completed for a particular client due to factors beyond this process's control, such as the client declined to participate in CSAT, the leadership team decided not to collect a CSAT from that client, or that the client was contacted multiple times in the attempt to collect CSAT but was unsuccessful. We still need a way to track that these clients have been reached out to and we haven't missed them in our process. We want to be able to track these "Not Applicable" records but not have them counted again the CSAT Score calculations nor reflected in the CSAT Report (Reports > CSAT) as a completed "CSAT".
Therefore, I request a new item added to the "Referral Score" drop-down menu on the CSAT tab in Project Details:
When user clicks "Add CSAT" and the Project CSAT Details box appears, "Not Applicable" should appear at the bottom (i.e., after "0" add "Not Applicable") of the "On the scale of 0-10, would you refer Logic20/20 to a friend or colleague?" dropdown.

Selection of the "Not Applicable" rating has the following impacts to the system:
The CSAT record with a "Not Applicable" rating should be displayed on the "CSAT" tab in Project Details. "Not Applicable" is displayed in the "Referral Score" column.
On the CSAT report, any records with a "Referral Score" value of "Not Applicable" WILL NOT display in the CSAT report. Projects with a "Not Applicable" CSAT record will be excluded from the CSAT score calculation.
In the Excel Export, records with a "Not Applicable" CSAT record should be reflected in the spreadsheet.