Comments - Defect Number: 3196; Defect RecordID: 3192; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Entry > User unable to add time; Defect Description: Siddharth Sugathan terminated as a W2-Salary employee on 9/27/2013 and was reactivated with the new pay type of 1099 Hourly, effective 10/1/2013. Persons Details accurately reflects Siddharth's new hire date of 10/1/2013.
Siddharth is experiencing the issue that he cannot enter time into Time Entry before 10/20/2013 (see screenshot below). Could you please investigate and implement the appropriate fix?

Separate (not-high priority) request: In the lower right-hand side of the screenshot below, the Summary of Hours per day is not displaying properly (the column headings are shifted to the right of the hours). I have noticed this in other instances of entering time, and it occurs upon saving time. After this original issue is solved, I will de-prioritize the priority of this request from hotfix and ask for this portion of my request to be worked on accordingly.