Comments - Defect Number: 3212; Defect RecordID: 3208; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Default career counselor functionality; Defect Description: The message in Configuration>Company Resources>Default Career Counselor page reads that
Once selected, the Default Career Counselor is the person that all users in PM will default to in the event that they are either not assigned a Career Counselor or if their assigned Career Counselor is made Inactive or Terminated.

According to this when a default career counselor is selected then He will be set as career counselor if a person is not assigned with Career Counselor or if their assigned Career Counselor is made Inactive or Terminated.

But the current functionality in the system behaves as:

When a career counselor is terminated, the system assigns the person who satisfies all below conditions as the career counselor for his/her subordinates:
1. Practice area of the person is same as the terminated career counselor.
2. Persons status= Active
3. Persons seniority value <= 65

If multiple people satisfy the above criteria then the person with maximum personID value will be selected.

If there are no people who satisfy the above criteria then default career counselor (if he is with active status) will be set as career counselor for subordinates of terminated counselor. If the default career counselor is not with active status then the terminated persons practice area manager will be set as career counselor for subordinates.

Could you please let us know if this functionality is expected?

LN: I've noticed this too, and it was on my list to raise as an issue. Thank you for being proactive and doing it ahead of me. The current functionality is not performing as expected and should be fixed. I expect the functionality to work as described as you mention on the Career Counselor page that there is a default value that remains consistent regardless of Practice Area or seniority as the current functionality is currently working. Can we add an "Unassigned" value in the Career Counselor drop-down menu? It would be the default value when a new Person record is created. If the individual's assigned Career Counselor's status is changed to "Inactive" or "Terminated," then the field would revert back to the "Unassigned" default.