Comments - Defect Number: 3221; Defect RecordID: 3217; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Consulting Demand report; Defect Description: Clicking on bar graph shows a transaction report in consulting demand report.
If the period filter is set to dates of current year then the transaction report is being generated correctly.
If the period filter is set to dates other than current year then in number of resources in graphs tab is displayed correctly but transaction report that is displayed by clicking on bar is being shown incorrectly.

For example follow the steps given below:

Steps to reproduce:

In project detail of a project:
1. Add a milestone with start date as 12/1/2013 and end date as 12/31/2013
2. Add two strawmen to that milestone with date as 12/1/2013 and end date as 12/31/2013

In Consulting Demand report:
1. Set Graph type= Resource View By Title (or) Resource View By Skill
2. Sales stage=active
3. Title=All titles
4. Period=custom dates (select dates from 12/1/2013-12/31/2013)
5. Click on Run report
6. Once the report is loaded, Go to graph tab
7. Verify if the number of resources is being displayed.
8. Click on the bar of graph this should result in transaction report of the bar

Actual:
Clicking on bar with given filters results in blank transaction report though there are resources needed. (Number of resources is being shown correctly in bar graph, but not in transaction report)


Expected:
Clicking on bar graph should result in transaction report of resources needed as given in picture.