Comments - Defect Number: 3222; Defect RecordID: 3218; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Person detail:Setting terminated Practice area owner as Career counselor.; Defect Description: Currently the system allows the practice area owner to be set as career counselor by clicking on Set to Practice Area Owner link. But, if the practice area has a terminated person as owner then by clicking on Set to Practice Area Owner link,he is set as career counselor.
Steps to reproduce:
1. Consider that Test, Person is set as practice area owner for Administration practice.
2. Terminate the practice areas owner.
3. Now in person detail page of some person who has practice area as Administration click on Set to Practice Area Owner and set Test, Person as Career counselor and save.Terminated person is set as career counselor

Could you please let us know if this scenario is expected?

LN: No, this scenario is not expected, so good catch. Can we implement a business rule that doesn't allow a terminated person to be set as a career counselor via the workflow you described above? Can the system automatically update the drop-down so that terminated people are not reflected? Or perhaps we need to create an "Unassigned" for "Practice Area Owner" like we are doing for Career Counselor so that when a Practice Area Owner is terminated, the system automatically updates the designation to "Unassigned". What is your suggestion?

[Harshitha Tenneti]
We recommend your last suggestion where we need to include a value Unassigned in Practice area owner drop down of Configuration > Company Resources > Practice Areas page, which is set on terminating a practice area owner.
For example,
Consider a person test, person who is assigned as an owner for administration practice area. If the person is terminated then Unassigned value will be set in Practice area owner field.
In person detail page of a person who has administration as practice area, when user clicks on link Set to Practice Area Owner then the career counselor is set as Unassigned.

Please let us know your response on this.